Add DNS creation to install
Script when using the "-s" parameter will also expect the FQDN of
the server (hostname.domain.com) to create a DNS entry for the
system.

./install -s host.domain.com

This will create a file called "my_hi2dns" in /usr/local/sbin.
Its contents can be added to a DNS servers entries for the Type 55
lookup on that DNS.