Update installer script - not tested
Cleaned up the installer script based on some found issues during
use.

The script now is more robust and has more options that are helpful
to creating HIs. This includes new command line arguments to
make it easier to perform.

Also restructured the file for better modularity and use to make
it more presentable and usable.

The major new options are:
-n; This is to create new HIP installs
-a; This is for appending new HIs to existing install